Fix for "set" command failure
Recently we have observed some failures as below:

  crash> set 2276866
  set: invalid kernel virtual address: 0  type: "stack contents"
  set: read of stack at 0 failed

  crash> ps 2276866
        PID    PPID  CPU       TASK        ST  %MEM      VSZ      RSS  COMM
    2276866 2276750  47  ff3a19fbd3c80000  ZO   0.0        0        0  sh

This is a regression issue that introduced by adding gdb stack unwind
support. When attempting to read from the stack, firstly, need to check
if the stack exists, otherwise it may fail in some corner cases. E.g:
there are some zombie processes(ZO) and the stack does not exist.
Furthermore this may also break the switching thread in gdb.

With the patch:
  crash> set 2276866
      PID: 2276866
  COMMAND: "sh"
     TASK: ff3a19fbd3c80000  [THREAD_INFO: ff3a19fbd3c80000]
      CPU: 47
    STATE: EXIT_DEAD|EXIT_ZOMBIE

Reported-by: Buland Kumar Singh <bsingh@redhat.com>
Signed-off-by: Lianbo Jiang <lijiang@redhat.com>