Fix prompt output interfering with piped commands in input files
When executing piped commands from input files, the prompt text was
incorrectly written to the pipe instead of stdout.
This caused commands read the prompt as their first line,
consuming it instead of the actual command output.

  crash> p cpu_info:0 | head -1
  per_cpu(cpu_info, 0) = $29 = {
  crash> p cpu_info:0 | head -2
  per_cpu(cpu_info, 0) = $4 = {
    x86 = 0x6,
  crash> cat crashrc1
  p cpu_info:0 | head -1
  crash> < crashrc1
  crash> p cpu_info:0 | head -1

  crash> cat crashrc2
  p cpu_info:0 | head -2
  crash> < crashrc2
  crash> p cpu_info:0 | head -2
  per_cpu(cpu_info, 0) = $30 = {
  crash>

Fix: Output prompt to incoming_fp (stdout) instead of fp (which may be the pipe
after setup_redirect()), ensuring prompts display on terminal while command output
goes to the pipe as intended.

Signed-off-by: neilfsun <neilfsun@tencent.com>
Signed-off-by: Feng Sun <loyou85@gmail.com>